AprilTag target is an integer

+++ src/main/java/frc/robot/game/ReefAprilTag.java
@@ -8,11 +8,11 @@ public enum ReefAprilTag {
   TEN(11, 20, -129),
   TWELVE(10, 21, 180);
 
-  public final double red;
-  public final double blue;
+  public final int red;
+  public final int blue;
   public final double angle; // -180 to 180, from the driver station perspective
 
-  ReefAprilTag(double red, double blue, double angle) {
+  ReefAprilTag(int red, int blue, double angle) {
     this.red = red;
     this.blue = blue;
     this.angle = angle;
